separate embarkjs in to specific classes related to a particular component
