+++ src/browser.js
@@ -2,8 +2,8 @@ import _EmbarkJS from './embark';
 
 var EmbarkJS = Object.assign({}, _EmbarkJS);
 
-EmbarkJS.checkWeb3 = function () {
-    _EmbarkJS.checkWeb3.call(this);
+EmbarkJS.Contract.checkWeb3 = function () {
+    _EmbarkJS.Contract.checkWeb3.call(this);
     if (!this.web3 && typeof (web3) !== 'undefined') {
         this.web3 = web3;
     } else if (!this.web3) {
