+++ src/messages.js
@@ -1,0 +1,39 @@
+let Messages = {};
+
+Messages.Providers = {};
+
+Messages.registerProvider = function (providerName, obj) {
+  this.Providers[providerName] = obj;
+};
+
+Messages.setProvider = function (provider, options) {
+  let providerObj = this.Providers[provider];
+
+  if (!providerObj) {
+    throw new Error('Unknown messages provider');
+  }
+
+  this.currentMessages = providerObj;
+
+  return providerObj.setProvider(options);
+};
+
+Messages.isAvailable = function () {
+  return this.currentMessages.isAvailable();
+};
+
+Messages.sendMessage = function (options) {
+  if (!this.currentMessages) {
+    throw new Error('Messages provider not set; e.g EmbarkJS.Messages.setProvider("whisper")');
+  }
+  return this.currentMessages.sendMessage(options);
+};
+
+Messages.listenTo = function (options, callback) {
+  if (!this.currentMessages) {
+    throw new Error('Messages provider not set; e.g EmbarkJS.Messages.setProvider("whisper")');
+  }
+  return this.currentMessages.listenTo(options, callback);
+};
+
+export default Messages;
